+++ packages/manager/src/features/LoadBalancers/LoadBalancerDetail/Configurations/constants.tsx
@@ -29,13 +29,12 @@ export const initialValues = {
 
 export const CONFIGURATION_COPY = {
   Certificates:
-    'TLS termination certificates create an encrypted link between your clients and Global Load Balancer and terminate incoming traffic on the load balancer. Once the load balancing policy is applied, traffic is forwarded to your service targets over encrypted TLS connections. Responses from your service targets to your clients are also encrypted.',
+    'TLS termination certificates create an encrypted link between your clients and Global Load Balancer, and terminate incoming traffic on the load balancer. Once the load balancing policy is applied, traffic is forwarded to your service targets over encrypted TLS connections. Responses from your service targets to your clients are also encrypted.',
   Port: 'The inbound port that the load balancer listens on.',
   Protocol: (
     <Typography>
       Set to either TCP, HTTP, or HTTPS. See{' '}
-      <Link to={AGLB_DOCS.Protocols}>Available Protocols</Link> for for
-      information.
+      <Link to={AGLB_DOCS.Protocols}>Available Protocols</Link> for information.
     </Typography>
   ),
 };
